Put classes in individual files
Breakout ComponentDesigner, BehaviorService, and ControlDesigner nested classes into separate files.
