+++ src/System.Windows.Forms.Design/src/System/Windows/Forms/Design/Behavior/BehaviorService.cs
@@ -29,7 +29,7 @@ namespace System.Windows.Forms.Design.Behavior
     ///  smart tags etc... as well as control many of the design-time behaviors: dragging,
     ///  selection, snap lines, etc...
     /// </summary>
-    public sealed class BehaviorService : IDisposable
+    public sealed partial class BehaviorService : IDisposable
     {
         private readonly IServiceProvider _serviceProvider;             // standard service provider
         private readonly AdornerWindow _adornerWindow;                  // the transparent window all glyphs are drawn to
@@ -586,654 +586,6 @@ namespace System.Windows.Forms.Design.Behavior
             }
         }
 
-        /// <summary>
-        ///  The AdornerWindow is a transparent window that resides ontop of the Designer's Frame. This window is used
-        ///  by the BehaviorService to  intercept all messages.  It also serves as a unified canvas on which to paint Glyphs.
-        /// </summary>
-        private class AdornerWindow : Control
-        {
-            private readonly BehaviorService _behaviorService;//ptr back to BehaviorService
-            private Control _designerFrame;//the designer's frame
-            private static MouseHook s_mouseHook; // shared mouse hook
-            private static readonly List<AdornerWindow> s_adornerWindowList = new List<AdornerWindow>();
-            private bool _processingDrag; // is this particular window in a drag operation
-
-            /// <summary>
-            ///  Constructor that parents itself to the Designer Frame and hooks all
-            ///  necessary events.
-            /// </summary>
-            internal AdornerWindow(BehaviorService behaviorService, Control designerFrame)
-            {
-                _behaviorService = behaviorService;
-                _designerFrame = designerFrame;
-                Dock = DockStyle.Fill;
-                AllowDrop = true;
-                Text = "AdornerWindow";
-                SetStyle(ControlStyles.Opaque, true);
-            }
-
-            /// <summary>
-            ///  The key here is to set the appropriate TransparetWindow style.
-            /// </summary>
-            protected override CreateParams CreateParams
-            {
-                get
-                {
-                    CreateParams cp = base.CreateParams;
-                    cp.Style &= ~(int)(User32.WS.CLIPCHILDREN | User32.WS.CLIPSIBLINGS);
-                    cp.ExStyle |= (int)User32.WS_EX.TRANSPARENT;
-                    return cp;
-                }
-            }
-
-            internal bool ProcessingDrag
-            {
-                get => _processingDrag;
-                set => _processingDrag = value;
-            }
-
-            /// <summary>
-            ///  We'll use CreateHandle as our notification for creating our mouse attacher.
-            /// </summary>
-            protected override void OnHandleCreated(EventArgs e)
-            {
-                base.OnHandleCreated(e);
-                s_adornerWindowList.Add(this);
-                if (s_mouseHook is null)
-                {
-                    s_mouseHook = new MouseHook();
-                }
-            }
-
-            /// <summary>
-            ///  Unhook and null out our mouseHook.
-            /// </summary>
-            protected override void OnHandleDestroyed(EventArgs e)
-            {
-                s_adornerWindowList.Remove(this);
-                // unregister the mouse hook once all adorner windows have been disposed.
-                if (s_adornerWindowList.Count == 0 && s_mouseHook != null)
-                {
-                    s_mouseHook.Dispose();
-                    s_mouseHook = null;
-                }
-                base.OnHandleDestroyed(e);
-            }
-
-            /// <summary>
-            ///  Null out our mouseHook and unhook any events.
-            /// </summary>
-            protected override void Dispose(bool disposing)
-            {
-                if (disposing)
-                {
-                    if (_designerFrame != null)
-                    {
-                        _designerFrame = null;
-                    }
-                }
-                base.Dispose(disposing);
-            }
-
-            internal Control DesignerFrame
-            {
-                get => _designerFrame;
-            }
-
-            /// <summary>
-            ///  Returns the display rectangle for the adorner window
-            /// </summary>
-            internal Rectangle DesignerFrameDisplayRectangle
-            {
-                get
-                {
-                    if (DesignerFrameValid)
-                    {
-                        return ((DesignerFrame)_designerFrame).DisplayRectangle;
-                    }
-                    else
-                    {
-                        return Rectangle.Empty;
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Returns true if the DesignerFrame is created and not being disposed.
-            /// </summary>
-            internal bool DesignerFrameValid
-            {
-                get
-                {
-                    if (_designerFrame is null || _designerFrame.IsDisposed || !_designerFrame.IsHandleCreated)
-                    {
-                        return false;
-                    }
-                    return true;
-                }
-            }
-
-            public IEnumerable<Adorner> Adorners { get; private set; }
-
-            /// <summary>
-            ///  Ultimately called by ControlDesigner when it receives a DragDrop message - here, we'll exit from 'drag mode'.
-            /// </summary>
-            internal void EndDragNotification()
-            {
-                ProcessingDrag = false;
-            }
-
-            /// <summary>
-            ///  Invalidates the transparent AdornerWindow by asking the Designer Frame beneath it to invalidate.
-            ///  Note the they use of the .Update() call for perf. purposes.
-            /// </summary>
-            internal void InvalidateAdornerWindow()
-            {
-                if (DesignerFrameValid)
-                {
-                    _designerFrame.Invalidate(true);
-                    _designerFrame.Update();
-                }
-            }
-
-            /// <summary>
-            ///  Invalidates the transparent AdornerWindow by asking the Designer Frame beneath it to invalidate.
-            ///  Note the they use of the .Update() call for perf. purposes.
-            /// </summary>
-            internal void InvalidateAdornerWindow(Region region)
-            {
-                if (DesignerFrameValid)
-                {
-                    //translate for non-zero scroll positions
-                    Point scrollPosition = ((DesignerFrame)_designerFrame).AutoScrollPosition;
-                    region.Translate(scrollPosition.X, scrollPosition.Y);
-
-                    _designerFrame.Invalidate(region, true);
-                    _designerFrame.Update();
-                }
-            }
-
-            /// <summary>
-            ///  Invalidates the transparent AdornerWindow by asking the Designer Frame beneath it to invalidate.
-            ///  Note the they use of the .Update() call for perf. purposes.
-            /// </summary>
-            internal void InvalidateAdornerWindow(Rectangle rectangle)
-            {
-                if (DesignerFrameValid)
-                {
-                    //translate for non-zero scroll positions
-                    Point scrollPosition = ((DesignerFrame)_designerFrame).AutoScrollPosition;
-                    rectangle.Offset(scrollPosition.X, scrollPosition.Y);
-
-                    _designerFrame.Invalidate(rectangle, true);
-                    _designerFrame.Update();
-                }
-            }
-
-            /// <summary>
-            ///  The AdornerWindow hooks all Drag/Drop notification so  they can be forwarded to the appropriate
-            ///  Behavior via the BehaviorService.
-            /// </summary>
-            protected override void OnDragDrop(DragEventArgs e)
-            {
-                try
-                {
-                    _behaviorService.OnDragDrop(e);
-                }
-                finally
-                {
-                    ProcessingDrag = false;
-                }
-            }
-
-            internal void EnableAllAdorners(bool enabled)
-            {
-                foreach (Adorner adorner in Adorners)
-                {
-                    adorner.EnabledInternal = enabled;
-                }
-                Invalidate();
-            }
-
-            private static bool IsLocalDrag(DragEventArgs e)
-            {
-                if (e.Data is DropSourceBehavior.BehaviorDataObject)
-                {
-                    return true;
-                }
-                else
-                {
-                    // Gets all the data formats and data conversion formats in the data object.
-                    string[] allFormats = e.Data.GetFormats();
-
-                    for (int i = 0; i < allFormats.Length; i++)
-                    {
-                        if (allFormats[i].Length == ToolboxFormat.Length &&
-                            string.Equals(ToolboxFormat, allFormats[i]))
-                        {
-                            return true;
-                        }
-                    }
-                }
-                return false;
-            }
-
-            /// <summary>
-            ///  The AdornerWindow hooks all Drag/Drop notification so  they can be forwarded to the appropriate
-            ///  Behavior via the BehaviorService.
-            /// </summary>
-            protected override void OnDragEnter(DragEventArgs e)
-            {
-                ProcessingDrag = true;
-
-                // determine if this is a local drag, if it is, do normal processing otherwise, force a
-                // PropagateHitTest.  We need to force this because the OLE D&D service suspends mouse messages
-                // when the drag is not local so the mouse hook never sees them.
-                if (!IsLocalDrag(e))
-                {
-                    _behaviorService._validDragArgs = e;
-                    User32.GetCursorPos(out Point pt);
-                    User32.MapWindowPoints(IntPtr.Zero, Handle, ref pt, 1);
-                    _behaviorService.PropagateHitTest(pt);
-                }
-                _behaviorService.OnDragEnter(null, e);
-            }
-
-            /// <summary>
-            ///  The AdornerWindow hooks all Drag/Drop notification so they can be forwarded to the appropriate
-            ///  Behavior via the BehaviorService.
-            /// </summary>
-            protected override void OnDragLeave(EventArgs e)
-            {
-                // set our dragArgs to null so we know not to send drag enter/leave events when we re-enter the dragging area
-                _behaviorService._validDragArgs = null;
-                try
-                {
-                    _behaviorService.OnDragLeave(null, e);
-                }
-                finally
-                {
-                    ProcessingDrag = false;
-                }
-            }
-
-            /// <summary>
-            ///  The AdornerWindow hooks all Drag/Drop notification so they can be forwarded to the appropriate
-            ///  Behavior via the BehaviorService.
-            /// </summary>
-            protected override void OnDragOver(DragEventArgs e)
-            {
-                ProcessingDrag = true;
-                if (!IsLocalDrag(e))
-                {
-                    _behaviorService._validDragArgs = e;
-                    User32.GetCursorPos(out Point pt);
-                    User32.MapWindowPoints(IntPtr.Zero, Handle, ref pt, 1);
-                    _behaviorService.PropagateHitTest(pt);
-                }
-
-                _behaviorService.OnDragOver(e);
-            }
-
-            /// <summary>
-            ///  The AdornerWindow hooks all Drag/Drop notification so they can be forwarded to the appropriate
-            ///  Behavior via the BehaviorService.
-            /// </summary>
-            protected override void OnGiveFeedback(GiveFeedbackEventArgs e)
-            {
-                _behaviorService.OnGiveFeedback(e);
-            }
-
-            /// <summary>
-            ///  The AdornerWindow hooks all Drag/Drop notification so they can be forwarded to the appropriate
-            ///  Behavior via the BehaviorService.
-            /// </summary>
-            protected override void OnQueryContinueDrag(QueryContinueDragEventArgs e)
-            {
-                _behaviorService.OnQueryContinueDrag(e);
-            }
-
-            /// <summary>
-            ///  Called by ControlDesigner when it receives a DragEnter message - we'll let listen to all Mouse
-            ///  Messages so we can send drag notifcations.
-            /// </summary>
-            internal void StartDragNotification()
-            {
-                ProcessingDrag = true;
-            }
-
-            /// <summary>
-            ///  The AdornerWindow intercepts all designer-related messages and forwards them to the BehaviorService
-            ///  for appropriate actions.  Note that Paint and HitTest messages are correctly parsed and translated to AdornerWindow coords.
-            /// </summary>
-            protected override void WndProc(ref Message m)
-            {
-                //special test hooks
-                if (m.Msg == (int)BehaviorService.WM_GETALLSNAPLINES)
-                {
-                    _behaviorService.TestHook_GetAllSnapLines(ref m);
-                }
-                else if (m.Msg == (int)BehaviorService.WM_GETRECENTSNAPLINES)
-                {
-                    _behaviorService.TestHook_GetRecentSnapLines(ref m);
-                }
-
-                switch ((User32.WM)m.Msg)
-                {
-                    case User32.WM.PAINT:
-                        {
-                            // Stash off the region we have to update.
-                            using var hrgn = new Gdi32.RegionScope(0, 0, 0, 0);
-                            User32.GetUpdateRgn(m.HWnd, hrgn, BOOL.TRUE);
-
-                            // The region we have to update in terms of the smallest rectangle that completely encloses
-                            // the update region of the window gives us the clip rectangle.
-                            RECT clip = new RECT();
-                            User32.GetUpdateRect(m.HWnd, ref clip, BOOL.TRUE);
-                            Rectangle paintRect = clip;
-
-                            using Region region = hrgn.CreateGdiPlusRegion();
-
-                            // Call the base class to do its painting.
-                            DefWndProc(ref m);
-
-                            // Now do our own painting.
-                            using Graphics g = Graphics.FromHwnd(m.HWnd);
-                            using PaintEventArgs pevent = new PaintEventArgs(g, paintRect);
-                            g.Clip = region;
-                            _behaviorService.PropagatePaint(pevent);
-
-                            break;
-                        }
-
-                    case User32.WM.NCHITTEST:
-                        Point pt = new Point((short)PARAM.LOWORD(m.LParam),
-                                             (short)PARAM.HIWORD(m.LParam));
-                        var pt1 = new Point();
-                        User32.MapWindowPoints(IntPtr.Zero, Handle, ref pt1, 1);
-                        pt.Offset(pt1.X, pt1.Y);
-                        if (_behaviorService.PropagateHitTest(pt) && !ProcessingDrag)
-                        {
-                            m.Result = (IntPtr)User32.HT.TRANSPARENT;
-                        }
-                        else
-                        {
-                            m.Result = (IntPtr)User32.HT.CLIENT;
-                        }
-                        break;
-
-                    case User32.WM.CAPTURECHANGED:
-                        base.WndProc(ref m);
-                        _behaviorService.OnLoseCapture();
-                        break;
-
-                    default:
-                        base.WndProc(ref m);
-                        break;
-                }
-            }
-
-            /// <summary>
-            ///  Called by our mouseHook when it spies a mouse message that the adornerWindow would be interested in.
-            ///  Returning 'true' signifies that the message was processed and should not continue to child windows.
-            /// </summary>
-            private bool WndProcProxy(ref Message m, int x, int y)
-            {
-                Point mouseLoc = new Point(x, y);
-                _behaviorService.PropagateHitTest(mouseLoc);
-                switch ((User32.WM)m.Msg)
-                {
-                    case User32.WM.LBUTTONDOWN:
-                        if (_behaviorService.OnMouseDown(MouseButtons.Left, mouseLoc))
-                        {
-                            return false;
-                        }
-                        break;
-
-                    case User32.WM.RBUTTONDOWN:
-                        if (_behaviorService.OnMouseDown(MouseButtons.Right, mouseLoc))
-                        {
-                            return false;
-                        }
-                        break;
-
-                    case User32.WM.MOUSEMOVE:
-                        if (_behaviorService.OnMouseMove(Control.MouseButtons, mouseLoc))
-                        {
-                            return false;
-                        }
-                        break;
-
-                    case User32.WM.LBUTTONUP:
-                        if (_behaviorService.OnMouseUp(MouseButtons.Left))
-                        {
-                            return false;
-                        }
-                        break;
-
-                    case User32.WM.RBUTTONUP:
-                        if (_behaviorService.OnMouseUp(MouseButtons.Right))
-                        {
-                            return false;
-                        }
-                        break;
-
-                    case User32.WM.MOUSEHOVER:
-                        if (_behaviorService.OnMouseHover(mouseLoc))
-                        {
-                            return false;
-                        }
-                        break;
-
-                    case User32.WM.LBUTTONDBLCLK:
-                        if (_behaviorService.OnMouseDoubleClick(MouseButtons.Left, mouseLoc))
-                        {
-                            return false;
-                        }
-                        break;
-
-                    case User32.WM.RBUTTONDBLCLK:
-                        if (_behaviorService.OnMouseDoubleClick(MouseButtons.Right, mouseLoc))
-                        {
-                            return false;
-                        }
-                        break;
-                }
-                return true;
-            }
-
-            /// <summary>
-            ///  This class knows how to hook all the messages to a given process/thread.
-            ///
-            ///  On any mouse clicks, it asks the designer what to do with the message, that is to eat it or propogate
-            ///  it to the control it was meant for.   This allows us to synchrounously process mouse messages when
-            ///  the AdornerWindow itself may be pumping messages.
-            /// </summary>
-            private class MouseHook
-            {
-                private AdornerWindow _currentAdornerWindow;
-                private uint _thisProcessID;
-                private GCHandle _mouseHookRoot;
-                private IntPtr _mouseHookHandle = IntPtr.Zero;
-                private bool _processingMessage;
-
-                private bool _isHooked; //VSWHIDBEY # 474112
-                private int _lastLButtonDownTimeStamp;
-
-                public MouseHook()
-                {
-#if DEBUG
-                    _callingStack = Environment.StackTrace;
-#endif
-                    HookMouse();
-                }
-#if DEBUG
-                readonly string _callingStack;
-                ~MouseHook()
-                {
-                    Debug.Assert(
-                        _mouseHookHandle == IntPtr.Zero,
-                        "Finalizing an active mouse hook.  This will crash the process.  Calling stack: " + _callingStack);
-                }
-#endif
-
-                public void Dispose()
-                {
-                    UnhookMouse();
-                }
-
-                private void HookMouse()
-                {
-                    Debug.Assert(s_adornerWindowList.Count > 0, "No AdornerWindow available to create the mouse hook");
-                    lock (this)
-                    {
-                        if (_mouseHookHandle != IntPtr.Zero || s_adornerWindowList.Count == 0)
-                        {
-                            return;
-                        }
-
-                        if (_thisProcessID == 0)
-                        {
-                            AdornerWindow adornerWindow = s_adornerWindowList[0];
-                            User32.GetWindowThreadProcessId(adornerWindow, out _thisProcessID);
-                        }
-
-                        var hook = new User32.HOOKPROC(MouseHookProc);
-                        _mouseHookRoot = GCHandle.Alloc(hook);
-                        _mouseHookHandle = User32.SetWindowsHookExW(
-                            User32.WH.MOUSE,
-                            hook,
-                            IntPtr.Zero,
-                            (uint)AppDomain.GetCurrentThreadId());
-                        if (_mouseHookHandle != IntPtr.Zero)
-                        {
-                            _isHooked = true;
-                        }
-                        Debug.Assert(_mouseHookHandle != IntPtr.Zero, "Failed to install mouse hook");
-                    }
-                }
-
-                private unsafe IntPtr MouseHookProc(User32.HC nCode, IntPtr wparam, IntPtr lparam)
-                {
-                    if (_isHooked && nCode == User32.HC.ACTION)
-                    {
-                        User32.MOUSEHOOKSTRUCT* mhs = (User32.MOUSEHOOKSTRUCT*)lparam;
-                        if (mhs != null)
-                        {
-                            try
-                            {
-                                if (ProcessMouseMessage(mhs->hWnd, unchecked((int)(long)wparam), mhs->pt.X, mhs->pt.Y))
-                                {
-                                    return (IntPtr)1;
-                                }
-                            }
-                            catch (Exception ex)
-                            {
-                                _currentAdornerWindow.Capture = false;
-                                if (ex != CheckoutException.Canceled)
-                                {
-                                    _currentAdornerWindow._behaviorService.ShowError(ex);
-                                }
-                                if (ClientUtils.IsCriticalException(ex))
-                                {
-                                    throw;
-                                }
-                            }
-                            finally
-                            {
-                                _currentAdornerWindow = null;
-                            }
-                        }
-                    }
-
-                    Debug.Assert(_isHooked, "How did we get here when we are diposed?");
-                    return User32.CallNextHookEx(new HandleRef(this, _mouseHookHandle), nCode, wparam, lparam);
-                }
-
-                private void UnhookMouse()
-                {
-                    lock (this)
-                    {
-                        if (_mouseHookHandle != IntPtr.Zero)
-                        {
-                            User32.UnhookWindowsHookEx(new HandleRef(this, _mouseHookHandle));
-                            _mouseHookRoot.Free();
-                            _mouseHookHandle = IntPtr.Zero;
-                            _isHooked = false;
-                        }
-                    }
-                }
-
-                private bool ProcessMouseMessage(IntPtr hWnd, int msg, int x, int y)
-                {
-                    if (_processingMessage)
-                    {
-                        return false;
-                    }
-
-                    foreach (AdornerWindow adornerWindow in AdornerWindow.s_adornerWindowList)
-                    {
-                        if (!adornerWindow.DesignerFrameValid)
-                        {
-                            continue;
-                        }
-
-                        _currentAdornerWindow = adornerWindow;
-                        IntPtr handle = adornerWindow.DesignerFrame.Handle;
-
-                        // if it's us or one of our children, just process as normal
-                        if (adornerWindow.ProcessingDrag || (hWnd != handle && User32.IsChild(new HandleRef(this, handle), hWnd).IsTrue()))
-                        {
-                            Debug.Assert(_thisProcessID != 0, "Didn't get our process id!");
-
-                            // Make sure the window is in our process
-                            User32.GetWindowThreadProcessId(hWnd, out uint pid);
-
-                            // If this isn't our process, bail
-                            if (pid != _thisProcessID)
-                            {
-                                return false;
-                            }
-
-                            try
-                            {
-                                _processingMessage = true;
-                                var pt = new Point(x, y);
-                                User32.MapWindowPoints(IntPtr.Zero, adornerWindow.Handle, ref pt, 1);
-                                Message m = Message.Create(hWnd, msg, (IntPtr)0, PARAM.FromLowHigh(pt.Y, pt.X));
-
-                                // No one knows why we get an extra click here from VS. As a workaround, we check the TimeStamp and discard it.
-                                if (m.Msg == (int)User32.WM.LBUTTONDOWN)
-                                {
-                                    _lastLButtonDownTimeStamp = User32.GetMessageTime();
-                                }
-                                else if (m.Msg == (int)User32.WM.LBUTTONDBLCLK)
-                                {
-                                    int lButtonDoubleClickTimeStamp = User32.GetMessageTime();
-                                    if (lButtonDoubleClickTimeStamp == _lastLButtonDownTimeStamp)
-                                    {
-                                        return true;
-                                    }
-                                }
-
-                                if (!adornerWindow.WndProcProxy(ref m, pt.X, pt.Y))
-                                {
-                                    // we did the work, stop the message propogation
-                                    return true;
-                                }
-                            }
-                            finally
-                            {
-                                _processingMessage = false;
-                            }
-                            break; // no need to enumerate the other adorner windows since only one can be focused at a time.
-                        }
-                    }
-                    return false;
-                }
-            }
-        }
-
         private bool PropagateHitTest(Point pt)
         {
             for (int i = _adorners.Count - 1; i >= 0; i--)
@@ -1285,76 +637,6 @@ namespace System.Windows.Forms.Design.Behavior
             // Returning false will cause the transparent window to return HTCLIENT when handling WM_NCHITTEST,
             // thus blocking underline window to receive mouse events.
             return true;
-        }
-
-        private class MenuCommandHandler : IMenuCommandService
-        {
-            private readonly BehaviorService _owner;            // ptr back to the behavior service
-            private readonly IMenuCommandService _menuService;  // core service used for most implementations of the IMCS interface
-            private readonly Stack<CommandID> _currentCommands = new Stack<CommandID>();
-
-            public MenuCommandHandler(BehaviorService owner, IMenuCommandService menuService)
-            {
-                _owner = owner;
-                _menuService = menuService;
-            }
-
-            public IMenuCommandService MenuService
-            {
-                get => _menuService;
-            }
-
-            void IMenuCommandService.AddCommand(MenuCommand command)
-            {
-                _menuService.AddCommand(command);
-            }
-
-            void IMenuCommandService.RemoveVerb(DesignerVerb verb)
-            {
-                _menuService.RemoveVerb(verb);
-            }
-
-            void IMenuCommandService.RemoveCommand(MenuCommand command)
-            {
-                _menuService.RemoveCommand(command);
-            }
-
-            MenuCommand IMenuCommandService.FindCommand(CommandID commandID)
-            {
-                try
-                {
-                    if (_currentCommands.Contains(commandID))
-                    {
-                        return null;
-                    }
-                    _currentCommands.Push(commandID);
-                    return _owner.FindCommand(commandID, _menuService);
-                }
-                finally
-                {
-                    _currentCommands.Pop();
-                }
-            }
-
-            bool IMenuCommandService.GlobalInvoke(CommandID commandID)
-            {
-                return _menuService.GlobalInvoke(commandID);
-            }
-
-            void IMenuCommandService.ShowContextMenu(CommandID menuID, int x, int y)
-            {
-                _menuService.ShowContextMenu(menuID, x, y);
-            }
-
-            void IMenuCommandService.AddVerb(DesignerVerb verb)
-            {
-                _menuService.AddVerb(verb);
-            }
-
-            DesignerVerbCollection IMenuCommandService.Verbs
-            {
-                get => _menuService.Verbs;
-            }
         }
 
         internal void StartDragNotification()
