+++ src/System.Windows.Forms.Design/src/System/Windows/Forms/Design/ControlDesigner.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.CodeDom;
 using System.Collections;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -22,7 +21,7 @@ namespace System.Windows.Forms.Design
     /// <summary>
     ///  Provides a designer that can design components that extend Control.
     /// </summary>
-    public class ControlDesigner : ComponentDesigner
+    public partial class ControlDesigner : ComponentDesigner
     {
 #pragma warning disable IDE1006 // Naming Styles - Public API
         protected static readonly Point InvalidPoint = new Point(int.MinValue, int.MinValue);
@@ -552,11 +551,6 @@ namespace System.Windows.Forms.Design
                     }
                 }
             }
-        }
-
-        private interface IDesignerTarget : IDisposable
-        {
-            void DefWndProc(ref Message m);
         }
 
         private void DataSource_ComponentRemoved(object sender, ComponentEventArgs e)
@@ -2519,231 +2513,6 @@ namespace System.Windows.Forms.Design
             return PARAM.ToInt(pt.X, pt.Y);
         }
 
-        public class ControlDesignerAccessibleObject : AccessibleObject
-        {
-            private readonly ControlDesigner _designer;
-            private readonly Control _control;
-            private IDesignerHost _host;
-            private ISelectionService _selSvc;
-
-            public ControlDesignerAccessibleObject(ControlDesigner designer, Control control)
-            {
-                _designer = designer;
-                _control = control;
-            }
-
-            public override Rectangle Bounds
-            {
-                get => _control.AccessibilityObject.Bounds;
-            }
-
-            public override string Description
-            {
-                get => _control.AccessibilityObject.Description;
-            }
-
-            private IDesignerHost DesignerHost
-            {
-                get
-                {
-                    if (_host is null)
-                    {
-                        _host = (IDesignerHost)_designer.GetService(typeof(IDesignerHost));
-                    }
-                    return _host;
-                }
-            }
-
-            public override string DefaultAction
-            {
-                get => "";
-            }
-
-            public override string Name
-            {
-                get => _control.Name;
-            }
-
-            public override AccessibleObject Parent
-            {
-                get => _control.AccessibilityObject.Parent;
-            }
-
-            public override AccessibleRole Role
-            {
-                get => _control.AccessibilityObject.Role;
-            }
-
-            private ISelectionService SelectionService
-            {
-                get
-                {
-                    if (_selSvc is null)
-                    {
-                        _selSvc = (ISelectionService)_designer.GetService(typeof(ISelectionService));
-                    }
-                    return _selSvc;
-                }
-            }
-
-            public override AccessibleStates State
-            {
-                get
-                {
-                    AccessibleStates state = _control.AccessibilityObject.State;
-                    ISelectionService s = SelectionService;
-                    if (s != null)
-                    {
-                        if (s.GetComponentSelected(_control))
-                        {
-                            state |= AccessibleStates.Selected;
-                        }
-                        if (s.PrimarySelection == _control)
-                        {
-                            state |= AccessibleStates.Focused;
-                        }
-                    }
-                    return state;
-                }
-            }
-
-            public override string Value
-            {
-                get => _control.AccessibilityObject.Value;
-            }
-
-            public override AccessibleObject GetChild(int index)
-            {
-                Debug.WriteLineIf(
-                    CompModSwitches.MSAA.TraceInfo,
-                    $"ControlDesignerAccessibleObject.GetChild({index})");
-
-                if (_control.AccessibilityObject.GetChild(index) is Control.ControlAccessibleObject childAccObj)
-                {
-                    AccessibleObject cao = GetDesignerAccessibleObject(childAccObj);
-                    if (cao != null)
-                    {
-                        return cao;
-                    }
-                }
-                return _control.AccessibilityObject.GetChild(index);
-            }
-
-            public override int GetChildCount() => _control.AccessibilityObject.GetChildCount();
-
-            private AccessibleObject GetDesignerAccessibleObject(Control.ControlAccessibleObject cao)
-            {
-                if (cao is null)
-                {
-                    return null;
-                }
-                if (DesignerHost.GetDesigner(cao.Owner) is ControlDesigner ctlDesigner)
-                {
-                    return ctlDesigner.AccessibilityObject;
-                }
-                return null;
-            }
-
-            public override AccessibleObject GetFocused()
-            {
-                if ((State & AccessibleStates.Focused) != 0)
-                {
-                    return this;
-                }
-                return base.GetFocused();
-            }
-
-            public override AccessibleObject GetSelected()
-            {
-                if ((State & AccessibleStates.Selected) != 0)
-                {
-                    return this;
-                }
-                return base.GetFocused();
-            }
-
-            public override AccessibleObject HitTest(int x, int y) => _control.AccessibilityObject.HitTest(x, y);
-        }
-
-        /// <summary>
-        ///  This TransparentBehavior is associated with the BodyGlyph for this ControlDesigner.  When the
-        ///  BehaviorService hittests a glyph w/a TransparentBehavior, all messages will be passed through the
-        ///  BehaviorService directly to the ControlDesigner. During a Drag operation, when the BehaviorService hittests
-        /// </summary>
-        internal class TransparentBehavior : Behavior.Behavior
-        {
-            readonly ControlDesigner _designer;
-            Rectangle _controlRect = Rectangle.Empty;
-
-            /// <summary>
-            ///  Constructor that accepts the related ControlDesigner.
-            /// </summary>
-            internal TransparentBehavior(ControlDesigner designer)
-            {
-                _designer = designer;
-            }
-
-            /// <summary>
-            ///  This property performs a hit test on the ControlDesigner to determine if the BodyGlyph should return
-            ///  '-1' for hit testing (letting all messages pass directly to the the control).
-            /// </summary>
-            internal bool IsTransparent(Point p) => _designer.GetHitTest(p);
-
-            /// <summary>
-            ///  Forwards DragDrop notification from the BehaviorService to the related ControlDesigner.
-            /// </summary>
-            public override void OnDragDrop(Glyph g, DragEventArgs e)
-            {
-                _controlRect = Rectangle.Empty;
-                _designer.OnDragDrop(e);
-            }
-
-            /// <summary>
-            ///  Forwards DragDrop notification from the BehaviorService to the related ControlDesigner.
-            /// </summary>
-            public override void OnDragEnter(Glyph g, DragEventArgs e)
-            {
-                if (_designer != null && _designer.Control != null)
-                {
-                    _controlRect = _designer.Control.RectangleToScreen(_designer.Control.ClientRectangle);
-                }
-                _designer.OnDragEnter(e);
-            }
-
-            /// <summary>
-            ///  Forwards DragDrop notification from the BehaviorService to the related ControlDesigner.
-            /// </summary>
-            public override void OnDragLeave(Glyph g, EventArgs e)
-            {
-                _controlRect = Rectangle.Empty;
-                _designer.OnDragLeave(e);
-            }
-
-            /// <summary>
-            ///  Forwards DragDrop notification from the BehaviorService to the related ControlDesigner.
-            /// </summary>
-            public override void OnDragOver(Glyph g, DragEventArgs e)
-            {
-                // If we are not over a valid drop area, then do not allow the drag/drop. Now that all
-                // dragging/dropping is done via the behavior service and adorner window, we have to do our own
-                // validation, and cannot rely on the OS to do it for us.
-                if (e != null && _controlRect != Rectangle.Empty && !_controlRect.Contains(new Point(e.X, e.Y)))
-                {
-                    e.Effect = DragDropEffects.None;
-                    return;
-                }
-                _designer.OnDragOver(e);
-            }
-
-            /// <summary>
-            ///  Forwards DragDrop notification from the BehaviorService to the related ControlDesigner.
-            /// </summary>
-            public override void OnGiveFeedback(Glyph g, GiveFeedbackEventArgs e)
-            {
-                _designer.OnGiveFeedback(e);
-            }
-        }
-
         internal void HookChildHandles(IntPtr firstChild)
         {
             IntPtr hwndChild = firstChild;
@@ -2819,222 +2588,11 @@ namespace System.Windows.Forms.Design
             HookChildControls(Control);
         }
 
-        private class ChildSubClass : NativeWindow, IDesignerTarget
-        {
-            private ControlDesigner _designer;
-
-            // AssignHandle calls NativeWindow::OnHandleChanged, but we do not override it so we should be okay
-            public ChildSubClass(ControlDesigner designer, IntPtr hwnd)
-            {
-                _designer = designer;
-                if (designer != null)
-                {
-                    designer.DisposingHandler += new EventHandler(OnDesignerDisposing);
-                }
-                AssignHandle(hwnd);
-            }
-
-            void IDesignerTarget.DefWndProc(ref Message m) => base.DefWndProc(ref m);
-
-            public void Dispose() => _designer = null;
-
-            private void OnDesignerDisposing(object sender, EventArgs e) => Dispose();
-
-            protected override void WndProc(ref Message m)
-            {
-                if (_designer is null)
-                {
-                    DefWndProc(ref m);
-                    return;
-                }
-
-                if (m.Msg == (int)User32.WM.DESTROY)
-                {
-                    _designer.RemoveSubclassedWindow(m.HWnd);
-                }
-                if (m.Msg == (int)User32.WM.PARENTNOTIFY && PARAM.LOWORD(m.WParam) == (short)User32.WM.CREATE)
-                {
-                    _designer.HookChildHandles(m.LParam); // they will get removed from the collection just above
-                }
-
-                // We want these messages to go through the designer's WndProc method, and we want people to be able
-                // to do default processing with the designer's DefWndProc.  So, we stuff ourselves into the designers
-                // window target and call their WndProc.
-                IDesignerTarget designerTarget = _designer.DesignerTarget;
-                _designer.DesignerTarget = this;
-                Debug.Assert(m.HWnd == Handle, "Message handle differs from target handle");
-                try
-                {
-                    _designer.WndProc(ref m);
-                }
-                catch (Exception ex)
-                {
-                    _designer.SetUnhandledException(Control.FromChildHandle(m.HWnd), ex);
-                }
-                finally
-                {
-                    // make sure the designer wasn't destroyed
-                    if (_designer != null && _designer.Component != null)
-                    {
-                        _designer.DesignerTarget = designerTarget;
-                    }
-                }
-            }
-        }
-
         internal void RemoveSubclassedWindow(IntPtr hwnd)
         {
             if (SubclassedChildWindows.ContainsKey(hwnd))
             {
                 SubclassedChildWindows.Remove(hwnd);
-            }
-        }
-
-        private class DesignerWindowTarget : IWindowTarget, IDesignerTarget, IDisposable
-        {
-            internal ControlDesigner _designer;
-            internal IWindowTarget _oldTarget;
-
-            public DesignerWindowTarget(ControlDesigner designer)
-            {
-                Control control = designer.Control;
-                _designer = designer;
-                _oldTarget = control.WindowTarget;
-                control.WindowTarget = this;
-            }
-
-            public void DefWndProc(ref Message m)
-            {
-                _oldTarget.OnMessage(ref m);
-            }
-
-            public void Dispose()
-            {
-                if (_designer != null)
-                {
-                    _designer.Control.WindowTarget = _oldTarget;
-                    _designer = null;
-                }
-            }
-
-            public void OnHandleChange(IntPtr newHandle)
-            {
-                _oldTarget.OnHandleChange(newHandle);
-                if (newHandle != IntPtr.Zero)
-                {
-                    _designer.OnHandleChange();
-                }
-            }
-
-            public void OnMessage(ref Message m)
-            {
-                // We want these messages to go through the designer's WndProc method, and we want people to be able
-                // to do default processing with the designer's DefWndProc.  So, we stuff ourselves into the designers
-                // window target and call their WndProc.
-                ControlDesigner currentDesigner = _designer;
-                if (currentDesigner != null)
-                {
-                    IDesignerTarget designerTarget = currentDesigner.DesignerTarget;
-                    currentDesigner.DesignerTarget = this;
-                    try
-                    {
-                        currentDesigner.WndProc(ref m);
-                    }
-                    catch (Exception ex)
-                    {
-                        currentDesigner.SetUnhandledException(currentDesigner.Control, ex);
-                    }
-                    finally
-                    {
-                        currentDesigner.DesignerTarget = designerTarget;
-                    }
-                }
-                else
-                {
-                    DefWndProc(ref m);
-                }
-            }
-        }
-
-        private class ChildWindowTarget : IWindowTarget, IDesignerTarget
-        {
-            private readonly ControlDesigner _designer;
-            private readonly Control _childControl;
-            private readonly IWindowTarget _oldWindowTarget;
-            private IntPtr _handle = IntPtr.Zero;
-
-            public ChildWindowTarget(ControlDesigner designer, Control childControl, IWindowTarget oldWindowTarget)
-            {
-                _designer = designer;
-                _childControl = childControl;
-                _oldWindowTarget = oldWindowTarget;
-            }
-
-            public IWindowTarget OldWindowTarget
-            {
-                get => _oldWindowTarget;
-            }
-
-            public void DefWndProc(ref Message m) => _oldWindowTarget.OnMessage(ref m);
-
-            public void Dispose()
-            {
-                // Do nothing. We will pick this up through a null DesignerTarget property when we come out of the message loop.
-            }
-
-            public void OnHandleChange(IntPtr newHandle)
-            {
-                _handle = newHandle;
-                _oldWindowTarget.OnHandleChange(newHandle);
-            }
-
-            public void OnMessage(ref Message m)
-            {
-                // If the designer has jumped ship, the continue partying on messages, but send them back to the original control.
-                if (_designer.Component is null)
-                {
-                    _oldWindowTarget.OnMessage(ref m);
-                    return;
-                }
-
-                // We want these messages to go through the designer's WndProc method, and we want people to be able
-                // to do default processing with the designer's DefWndProc.  So, we stuff the old window target into
-                // the designer's target and then call their WndProc.
-                IDesignerTarget designerTarget = _designer.DesignerTarget;
-                _designer.DesignerTarget = this;
-
-                try
-                {
-                    _designer.WndProc(ref m);
-                }
-                catch (Exception ex)
-                {
-                    _designer.SetUnhandledException(_childControl, ex);
-                }
-                finally
-                {
-                    // If the designer disposed us, then we should follow suit.
-                    if (_designer.DesignerTarget is null)
-                    {
-                        designerTarget.Dispose();
-                    }
-                    else
-                    {
-                        _designer.DesignerTarget = designerTarget;
-                    }
-
-                    // Controls (primarily RichEdit) will register themselves as drag-drop source/targets when they
-                    // are instantiated. Normally, when they are being designed, we will RevokeDragDrop() in their
-                    // designers. The problem occurs when these controls are inside a UserControl. At that time, we
-                    // do not have a designer for these controls, and they prevent the ParentControlDesigner's
-                    // drag-drop from working. What we do is to loop through all child controls that do not have a
-                    // designer (in HookChildControls()), and RevokeDragDrop() after their handles have been created.
-                    if (m.Msg == (int)User32.WM.CREATE)
-                    {
-                        Debug.Assert(_handle != IntPtr.Zero, "Handle for control not created");
-                        Ole32.RevokeDragDrop(_handle);
-                    }
-                }
             }
         }
 
@@ -3071,221 +2629,6 @@ namespace System.Windows.Forms.Design
                 }
                 Control.Invalidate(true);
             }
-        }
-
-        [ListBindable(false)]
-        [DesignerSerializer(typeof(DesignerControlCollectionCodeDomSerializer), typeof(CodeDomSerializer))]
-        internal class DesignerControlCollection : Control.ControlCollection, IList
-        {
-            readonly Control.ControlCollection _realCollection;
-            public DesignerControlCollection(Control owner) : base(owner)
-            {
-                _realCollection = owner.Controls;
-            }
-
-            public override int Count
-            {
-                get => _realCollection.Count;
-            }
-
-            object ICollection.SyncRoot
-            {
-                get => this;
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get => false;
-            }
-
-            bool IList.IsFixedSize
-            {
-                get => false;
-            }
-
-            public new bool IsReadOnly
-            {
-                get => _realCollection.IsReadOnly;
-            }
-
-            int IList.Add(object control) => ((IList)_realCollection).Add(control);
-
-            public override void Add(Control c) => _realCollection.Add(c);
-
-            public override void AddRange(Control[] controls) => _realCollection.AddRange(controls);
-
-            bool IList.Contains(object control) => ((IList)_realCollection).Contains(control);
-
-            public new void CopyTo(Array dest, int index) => _realCollection.CopyTo(dest, index);
-
-            public override bool Equals(object other) => _realCollection.Equals(other);
-
-            public new IEnumerator GetEnumerator() => _realCollection.GetEnumerator();
-
-            public override int GetHashCode() => _realCollection.GetHashCode();
-
-            int IList.IndexOf(object control) => ((IList)_realCollection).IndexOf(control);
-
-            void IList.Insert(int index, object value) => ((IList)_realCollection).Insert(index, value);
-
-            void IList.Remove(object control) => ((IList)_realCollection).Remove(control);
-
-            void IList.RemoveAt(int index) => ((IList)_realCollection).RemoveAt(index);
-
-            object IList.this[int index]
-            {
-                get => ((IList)_realCollection)[index];
-                set => throw new NotSupportedException();
-            }
-
-            public override int GetChildIndex(Control child, bool throwException) => _realCollection.GetChildIndex(child, throwException);
-
-            // we also need to redirect this guy
-            public override void SetChildIndex(Control child, int newIndex) => _realCollection.SetChildIndex(child, newIndex);
-
-            public override void Clear()
-            {
-                for (int i = _realCollection.Count - 1; i >= 0; i--)
-                {
-                    if (_realCollection[i] != null &&
-                        _realCollection[i].Site != null &&
-                        TypeDescriptor.GetAttributes(_realCollection[i]).Contains(InheritanceAttribute.NotInherited))
-                    {
-                        _realCollection.RemoveAt(i);
-                    }
-                }
-            }
-        }
-
-        // Custom code dom serializer for the DesignerControlCollection. We need this so we can filter out controls
-        // that aren't sited in the host's container.
-        internal class DesignerControlCollectionCodeDomSerializer : CollectionCodeDomSerializer
-        {
-            protected override object SerializeCollection(
-                IDesignerSerializationManager manager,
-                CodeExpression targetExpression,
-                Type targetType,
-                ICollection originalCollection,
-                ICollection valuesToSerialize)
-            {
-                ArrayList subset = new ArrayList();
-                if (valuesToSerialize != null && valuesToSerialize.Count > 0)
-                {
-                    foreach (object val in valuesToSerialize)
-                    {
-                        if (val is IComponent comp && comp.Site != null && !(comp.Site is INestedSite))
-                        {
-                            subset.Add(comp);
-                        }
-                    }
-                }
-                return base.SerializeCollection(manager, targetExpression, targetType, originalCollection, subset);
-            }
-        }
-
-        private class DockingActionList : DesignerActionList
-        {
-            private readonly ControlDesigner _designer;
-            private readonly IDesignerHost _host;
-
-            public DockingActionList(ControlDesigner owner) : base(owner.Component)
-            {
-                _designer = owner;
-                _host = GetService(typeof(IDesignerHost)) as IDesignerHost;
-            }
-
-            private string GetActionName()
-            {
-                PropertyDescriptor dockProp = TypeDescriptor.GetProperties(Component)["Dock"];
-                if (dockProp != null)
-                {
-                    DockStyle dockStyle = (DockStyle)dockProp.GetValue(Component);
-                    if (dockStyle == DockStyle.Fill)
-                    {
-                        return SR.DesignerShortcutUndockInParent;
-                    }
-                    else
-                    {
-                        return SR.DesignerShortcutDockInParent;
-                    }
-                }
-                return null;
-            }
-
-            public override DesignerActionItemCollection GetSortedActionItems()
-            {
-                DesignerActionItemCollection items = new DesignerActionItemCollection();
-                string actionName = GetActionName();
-                if (actionName != null)
-                {
-                    items.Add(new DesignerActionVerbItem(new DesignerVerb(GetActionName(), OnDockActionClick)));
-                }
-                return items;
-            }
-
-            private void OnDockActionClick(object sender, EventArgs e)
-            {
-                if (sender is DesignerVerb designerVerb && _host != null)
-                {
-                    using (DesignerTransaction t = _host.CreateTransaction(designerVerb.Text))
-                    {
-                        // set the dock prop to DockStyle.Fill
-                        PropertyDescriptor dockProp = TypeDescriptor.GetProperties(Component)["Dock"];
-                        DockStyle dockStyle = (DockStyle)dockProp.GetValue(Component);
-                        if (dockStyle == DockStyle.Fill)
-                        {
-                            dockProp.SetValue(Component, DockStyle.None);
-                        }
-                        else
-                        {
-                            dockProp.SetValue(Component, DockStyle.Fill);
-                        }
-                        t.Commit();
-                    }
-                }
-            }
-        }
-
-        private class CanResetSizePropertyDescriptor : PropertyDescriptor
-        {
-            private readonly PropertyDescriptor _basePropDesc;
-
-            public CanResetSizePropertyDescriptor(PropertyDescriptor pd) : base(pd)
-            {
-                _basePropDesc = pd;
-            }
-
-            public override Type ComponentType
-            {
-                get => _basePropDesc.ComponentType;
-            }
-
-            public override string DisplayName
-            {
-                get => _basePropDesc.DisplayName;
-            }
-
-            public override bool IsReadOnly
-            {
-                get => _basePropDesc.IsReadOnly;
-            }
-
-            public override Type PropertyType
-            {
-                get => _basePropDesc.PropertyType;
-            }
-
-            // since we can't get to the DefaultSize property, we use the existing ShouldSerialize logic.
-            public override bool CanResetValue(object component) => _basePropDesc.ShouldSerializeValue(component);
-
-            public override object GetValue(object component) => _basePropDesc.GetValue(component);
-
-            public override void ResetValue(object component) => _basePropDesc.ResetValue(component);
-
-            public override void SetValue(object component, object value) => _basePropDesc.SetValue(component, value);
-
-            // we always want to serialize values.
-            public override bool ShouldSerializeValue(object component) => true;
         }
     }
 }
